### 2017-02-04 #178 @whedon
[PRE REVIEW]: skedm: Empirical Dynamic Modeling
**Submitting author:** @nickc1 (<a href="http://orcid.org/0000-0003-2262-663X">Nicholas Cortale</a>)
**Repository:** <a href="https://github.com/NickC1/skedm" target ="_blank">https://github.com/NickC1/skedm</a>
**Version:** v0.0.1
