### 2017-02-04 #178 @arfon
👋 @jakevdp - this looks like it might be right up your street. Would you be willing to edit this submission for JOSS?
