### 2017-02-24 8255a32
fixed spelling in readme
