+++ README.md
@@ -2,8 +2,8 @@
 skedm
 =================================
 
-[![build status](https://travis-ci.org/NickC1/skedm.svg?branch=master)](https://travis-ci.org/NickC1/skedm)
-[![license](https://img.shields.io/github/license/mashape/apistatus.svg)](https://github.com/NickC1/skedm/blob/master/LICENSE)
+[![build status](https://travis-ci.org/nickc1/skedm.svg?branch=master)](https://travis-ci.org/nickc1/skedm)
+[![license](https://img.shields.io/github/license/mashape/apistatus.svg)](https://github.com/nickc1/skedm/blob/master/LICENSE)
 
 
 [DOCUMENTATION][read-the-docs]
