### 2017-03-14 #178 @jakevdp
Thanks to @johnh2o2 for being willing to review!
