### 2017-03-14 #207 @whedon
[REVIEW]: skedm: Empirical Dynamic Modeling
**Submitting author:** @nickc1 (<a href="http://orcid.org/0000-0003-2262-663X">Nicholas Cortale</a>)
**Repository:** <a href="https://github.com/NickC1/skedm" target ="_blank">https://github.com/NickC1/skedm</a>
**Version:** v0.0.1
**Archive:** <a href="http://dx.doi.org/10.6084/m9.figshare.4887398.v1" target="_blank">10.6084/m9.figshare.4887398.v1</a>
