### 2017-03-17 nickc1/skedm#1 @johnh2o2
BUG: the .ipynb in the scripts directory throws two exceptions
When running `skedm_examples.ipynb`, everything ran fine except for two `AttributeError`s, which were thrown because of misnamed functions:

* In 2D section: `X = data.chaos2D(sz=256)` throws `AttributeError` (`module skedm.data has no attribute chaos2D`) -- should be `chaos_2d` and not `chaos2D`
* Attribute error: `module 'skedm.data' has no attribute 'voronoiMatrix'` (should be `voronoi_matrix`).
