### 2017-03-17 nickc1/skedm#2 @johnh2o2
Version: Does the release version given match the GitHub release (v0.0.1)?
The release version given is v0.0.1, but the version for `skedm` is `0.2.0`.
