### 2017-03-17 #207 @johnh2o2
Hi @nickc1,

I've raised two issues (about the [version](https://github.com/nickc1/skedm/issues/2) and two [bugs in the ipynb](https://github.com/nickc1/skedm/issues/1)) in your GitHub repository. The rest of my review is below. Let me know if you have any questions or comments regarding anything!

## Overall
I'm not sure if I need to officially assign some status for this review, but currently I would say this is extremely close to submission with very minor issues. The vast majority of issues are with documentation, which is already quite thorough.

## Functionality
* The example `.ipynb` provides a large variety of both 1D and 2D examples with visualizations, all of which successfully execute (after fixing two typos).
* Unit tests of core functionality provide additional assurances of functionality.

## Performance
* There are only a few cursory and casual benchmarks, all of which were consistent with timing when run on my laptop. It (optionally) may be useful to present some basic timing analysis, but unless there is a discrepancy between stated and actual performance, I don't see this as a problem.

## API documentation
* Generally very good. Plots, in particular, were particularly helpful in illustrating some otherwise difficult to visualize points.
* Some detail is missing or glossed over that if provided should help clarify things for the reader that stumbles across your code. (I've detailed these below)
* In your docstrings for the `Classification` and `Regression` classes, I would try to add some more detail in the header
    * For instance Classification just says "Nonlinear classification", but nonlinear classification of what?
    * It would help to give some examples in the docstring (which you have for several functions):
    ```python
    """
    ...
    Examples
    --------
    >>> classifier = skedm.skedm.Classification(weights='distance')
    >>> classifier.fit(Xtrain, ytrain)
    ...
    """
    ```
### Install
* Looks good

### Quick example
* Great introduction!
* Great explanation of how to pick the time lag (mutual information minimum)
* This code I have trouble understanding
    ```python
    M.fit(Xtrain, ytrain) #fit the data (rebuilding the attractor)

    nn_list = np.arange(1,200,dtype='int') #first 200 NN
    ypred = M.predict(Xtest, nn_list)

    score = M.score(ytest) #score
    ```

    * What exactly is `score`? (I wasn't able to find the answer in the Score section)
    * What does `nn_list` represent? If I run `M.predict(Xtest, [ 1 ])` does this try to predict `y_test` by using only the closest neighbor for each of the points in `Xtest`? If I run `M.predict(Xtest, [1, 3])` does this use only the closest and third closest neighbor?
* The plot underneath makes sense, but I would explicitly show either some code that produces it or mention that this is generated by performing the same embed/regress/score procedure above, but for various `predict` and `nn_list` values (which correspond to the Forecast Distance and Near Neighbors).
* When talking about the forecast skill of each near neighbor, are you talking about how well you do by using a single neighbor trajectory? And you can choose to use the Nth nearest neighbor? If so, I think the X label on the plot underneath should be "Nth nearest neighbor" or something to denote that you aren't aggregating N near neighbors.

### Generate Data
* Some parameters are missing or inconsistent with the equation given in the docstring
    * Example: logistic map: `X(t+1) = rX(t)(1-X(t)) + random`
        * However, parameters are `A`, and `noise` should probably be used
    * `noisy_periodic`: use noise, freq in equation, etc.

### Embed
* You have `X, y = E.embed_vectors_1d(lag, emb, predict)`, but I think it might help to explicitly explain what `X` and `y` are via explaination in the text or via denoting them on your plots. I assume the green values are `y[n]` (vector of length `predict`) and the red values are `X[n]` (vector of length `embed`)?

### Predict
* Good overview.

### Score
* This is where I'm most confused -- I didn't see any explicit definition for what exactly
    `score` was; is this the $$R^2$$ correlation coefficient? Or something else?
* "After calculating near neighbors, calculating the forecast and forecast skill, the following plot is produced."
    * Too vague -- the plot was confusing to me. Are you plotting the "score" for predictions done using single neighbors? Meaning, do you say "ok lets use the 1st-closest neighbor", then compute R^2 and the average distance away that the 1st nearest neighbor is? Then moving on to the 2nd-closest, etc.?
    * The sentence below the plot adds some clarification, but it would be helpful to have some verbose code snippets that explain exactly what is being done to produce the plot

## Paper
* Great! (The references section should be blank, right?)
