### 2017-03-18 #207 @arfon
> Great! (The references section should be blank, right?)

Yep, these will be added at paper compilation time (by Pandoc)
