### 2017-03-20 96ca941
fixed typos in notebook
