### 2017-03-21 b063c5c
updated docstrings with examples and clarified the embedding docstrings
