+++ skedm/skedm.py
@@ -1,8 +1,5 @@
-"""
-Scikit learn implementation of nonlinear forecasting.
-
-By Nick Cortale
-"""
+#Scikit Empirical Dynamic Modeling
+# By Nick Cortale
 
 from . import utilities
 from scipy import stats as stats
@@ -11,8 +8,9 @@ import numpy as np
 from sklearn import metrics as skmetrics
 
 class Regression:
-    """
-    Nonlinear Regression
+    """Regression using a k-nearest neighbors method. Predictions can be made
+    for each nearest neighbor (`predict_individual`) or by averaging the k
+    nearest neighbors (`predict`).
 
     Parameters
     ----------
@@ -21,6 +19,17 @@ class Regression:
 
             - 'uniform' : uniform weighting
             - 'distance' : weighted as 1/distance
+
+    Example
+    -------
+    >>> X # embed time series of shape (nsamples, embedding dimension)
+    >>> y # future trajectory of a point of shape (nsamples, num predictions)
+    >>> import skedm as edm
+    >>> R = edm.Regression()
+    >>> train_len = int(len(X)*.75) # train on 75 percent
+    >>> R.fit(X[0:train_len], y[0:train_len])
+    >>> preds = R.predict(X[train_len:], [0,10,20]) # test at 1, 10, and 20 nn
+    >>> score = M.score(ytest) # Calculate coefficient of determination
     """
 
     def __init__(self,weights='uniform'):
@@ -34,9 +43,9 @@ class Regression:
         Parameters
         ----------
         Xtrain : 2D array
-            Features (nsamples,nfeatures).
+            Embed training time series. Features shape (nsamples, nfeatures).
         ytrain : 2D array
-            Targets (nsamples,ntargets).
+            Future trajectory of the points. Targets Shape (nsamples,ntargets).
 
         """
 
@@ -108,7 +117,7 @@ class Regression:
         return ypred
 
 
-    def score(self,ytest,how='score'):
+    def score(self, ytest, how='score'):
         """Score the predictions.
 
         Parameters
@@ -201,8 +210,9 @@ class Regression:
         return mean, std
 
 class Classification:
-    """
-    Nonlinear classification.
+    """Classification using a k-nearest neighbors method. Predictions can be
+    made for each nearest neighbor (`predict_individual`) or by averaging the k
+    nearest neighbors (`predict`).
 
     Parameters
     ----------
@@ -211,6 +221,17 @@ class Classification:
 
         - 'uniform' : uniform weighting
         - 'distance' : weighted as 1/distance
+
+    Example
+    -------
+    >>> X # embed series of shape (nsamples, embedding dimension)
+    >>> y # future trajectory of a point of shape (nsamples, num predictions)
+    >>> import skedm as edm
+    >>> R = edm.Classification()
+    >>> train_len = int(len(X)*.75) # train on 75 percent
+    >>> R.fit(X[0:train_len], y[0:train_len])
+    >>> preds = R.predict(X[train_len:], [0,10,20]) # test at 1, 10, and 20 nn
+    >>> score = M.score(ytest) # Calculate klecka's tau
     """
 
     def __init__(self,weights='uniform'):
@@ -261,6 +282,11 @@ class Classification:
         nn_list : 1d array of ints
             Neighbors to be tested.
 
+        Returns
+        -------
+        Ypred : list
+            Predictions returned for each nn value in nn_list. It is the same
+            length as nn_list.
         """
 
         self.dist_calc(Xtest)
@@ -311,6 +337,11 @@ class Classification:
         nn_list : 1d array of ints
             Neighbors to be tested.
 
+        Returns
+        -------
+        Ypred : list
+            Predictions returned for each nn value in nn_list. It is the same
+            length as nn_list.
         """
 
[…]
