### 2017-03-22 732ad6b
updated the Quick Example sections of the docs to improve clarity of input/output and added anotherreference for redm that was recently published
