+++ paper/paper.bib
@@ -8,6 +8,19 @@
   year={1999}
 }
 
+@misc{hao_ye_2017_376663,
+  author       = {Hao Ye and
+                  Jun Cai and
+                  Adam Clark and
+                  janecowles and
+                  Oliver Keyes and
+                  Ethan White},
+  title        = {ha0ye/rEDM: added readme},
+  month        = mar,
+  year         = 2017,
+  doi          = {10.5281/zenodo.376663},
+  url          = {https://doi.org/10.5281/zenodo.376663}
+}
 
 @article{scikit-learn,
   title={Scikit-learn: Machine learning in Python},
