### 2017-03-22 325243a
Cleaned the docstrings for data.py. Now reflects the correct parameters
