### 2017-03-22 f500188
updated Embed documentation section
