### 2017-03-22 #207 @jakevdp
@johnh2o2 — thanks for the thorough review. I think that some of the confusion you noted could be remedied by somewhere pointing to the scikit-learn API, which this seems to be modeled after.

@nickc1 — I think these are a nice set of comments and suggestions, and will improve the usability of your code. I hope you're able to address them!
