### 2017-03-22 #207 @nickc1
@jakevdp -- I definitely agree. Working down the list as I type this.

@johnh2o2 -- Appreciate the thorough comments. It's great getting an outside perspective on something that has been only used in our lab.  Hoping to have it wrapped up by the end of today! Thanks again.
