### 2017-03-22 1617264
clarified the score section with new plots
