### 2017-03-22 1d789cf
finished the scoring section
