### 2017-03-23 ccff3ed
Added some code to the performance section to clarify what I was testing
