### 2017-03-23 60ecdd0
fixed some docstring formatting errors for the embed class
