### 2017-03-23 4873ca5
fixed some typos
