### 2017-03-23 #207 @nickc1
Hey John @johnh2o2 ,

Thanks for the helpful comments. They made the documentation a lot clearer. Also, sorry for the lag in response time. I'm changing jobs at the end of next week, so I am getting pulled in a lot of different directions. Thanks for taking the time to review this! Let me know if you have any questions or want me to point you to specific commits or need me to clarify some of the new material.

Thanks again,
Nick

## Performance

> It (optionally) may be useful to present some basic timing analysis, but unless there is a discrepancy between stated and actual performance, I don't see this as a problem.

Added a section on performance in the documentation to make this more clear/helpful.

## API documentation

> In your docstrings for the Classification and Regression classes, I would try to add some more detail in the header. For instance Classification just says "Nonlinear classification", but nonlinear classification of what?

Updated docstrings with examples and fixed some of the formatting issues in the docstrings of the Embed class.

## Quick example

> What exactly is score? (I wasn't able to find the answer in the Score section) What does nn_list represent? If I run M.predict(Xtest, [ 1 ]) does this try to predict y_test by using only the closest neighbor for each of the points in Xtest? If I run M.predict(Xtest, [1, 3]) does this use only the closest and third closest neighbor?

I've added some new plots that (I think) do a better job of what the score represents. I also overhauled the score section to make it more clear to the reader what is actually going on. This comment was invaluable as I never really explained what score actually was. Appreciate this.

> When talking about the forecast skill of each near neighbor, are you talking about how well you do by using a single neighbor trajectory? And you can choose to use the Nth nearest neighbor? If so, I think the X label on the plot underneath should be "Nth nearest neighbor" or something to denote that you aren't aggregating N near neighbors.

Added an example of each in the quick example section to make this more clear.

## Generate Data

> Some parameters are missing or inconsistent with the equation given in the docstring Example: logistic map: X(t+1) = rX(t)(1-X(t)) + random

Updated the docstrings to represent the correct parameters.

## Embed

> You have X, y = E.embed_vectors_1d(lag, emb, predict), but I think it might help to explicitly explain what X and y are via explanation in the text or via denoting them on your plots. I assume the green values are y[n] (vector of length predict) and the red values are X[n] (vector of length embed)?

Explicitly stated now.

## Score

> This is where I'm most confused -- I didn't see any explicit definition for what exactly
score was; is this the $$R^2$$ correlation coefficient? Or something else?

I've completely overhauled this section as it was quite confusing. Let me know what you think.

## Paper

> Great! (The references section should be blank, right?)

I added a reference. An R package was actually released between the submission and now with similar functionality, so I decided to cite it in the paper.md.

## Statement of Need

Added a statement of need at the bottom of the documentation index.

## Versioning

If everything looks good, I'll change the version to 1.0 and push that to pypi.
