### 2017-03-24 nickc1/nonlinpy#1 @nickc1
Thanks John. These are resolved in the following commit:

https://github.com/nickc1/skedm/commit/b063c5c81ab48bba179991cccc933f10ef8695bf
