### 2017-03-24 #207 @johnh2o2
Hi @nickc1,

Things look much better! I have a few additional suggestions for you; once these are resolved I think everything should be ready to go.

## Version

> Version: Does the release version given match the GitHub release (v0.0.1)?

It's still `0.2.0` in the repo -- I would consider changing either the repo version, the Github release version (I'm not sure where that comes from) or both.

## Documentation

### Embed

> Here `X` is the red matrix (features) and `y` is the red matrix above (targets)

I think you mean `y`is the **green** matrix above (right?).

### Generate data
Better, however there is still some confusing/incomplete documentation:

* For the Lorenz attractor, if you have time/it's easy, it might make sense to let the user change the default `sigma`, `beta`, and `rho` values.
* For the `chaos_2d` and `chaos_3d` -- could you provide a link with more information about the logistic map diffused in space? Or if it's simple to give an explicit definition that would be helpful too
* Same with `periodic_2d` (i.e. what is the explicit form of the equation?)
* In `periodic_brown`, the equaiton given is `X = sin(y + noise*2pi)`, however there's a `freq` param; I'm assuming it's `X = sin(2 * pi * (freq * y + noise))`?
* For `skedm.data.noise_2d(sz=128, seed=36)`; I imagine some people will want to know what library/algorithm you're using as the random number generation; I would just mention what library/function you're using (e.g. numpy, etc.).
* `random_circles` -- what is a "random gaussian circle?"
* `circle_in_circle`, `circles_with_stuff`,`random_sized_circles`, I'm not sure what the form of the output is. What does the `blobber`/`blobber2` function do? If you're returning A 2-D array of size `(sz, sz)`, what does that array represent? Is it a bit-map image? The explanation for these should be a bit more specific
* In the `voronoi_diagram`, you say `sz -- row and column size of the space in which the circle is placed.`, but I think that's a typo since these are polygons. The return value is also unclear to me (see previous note)

Other than these suggestions, everything else looks good from here. The `ipynb` notebook runs all cells without raising any exceptions, so that looks good now.

Let me know if you have any objections or if anything isn't clear!

John
