### 2017-03-27 37d6f83
fixed a typo in the embedding documentation
