### 2017-03-29 6653052
updated docstrings for the data module and added a notebook showing examples of their output
