+++ skedm/data.py
@@ -1,21 +1,24 @@
 #
 # Generates data for skedm (SciKit Emperical Dynamic Modeling)
 #
-# Available data:
 # By Nick Cortale
-
-
 
 import numpy as np
 from numpy import genfromtxt
-from scipy import integrate
+#from scipy import integrate
 from sklearn import neighbors
+import scipy.ndimage
 
 
-def logistic_map(sz=256, A=3.99, seed=36, noise=0):
-    """Logistic map.
+def logistic_map(sz=256, A=3.99, seed=36, eta=0):
+    """Solutions to the logistic map for a given amount of time steps [1]_.
 
-    X(t+1) = AX(t)(1 - X(t)) + noise
+    .. math::
+
+      X_{t+1} = AX_t(1 - X_t) + \eta
+
+    where :math:`A` is the parameter that controls chaos, and :math:`\eta` is
+    the amplitude of the noise.
 
     Parameters
     ----------
@@ -25,14 +28,18 @@ def logistic_map(sz=256, A=3.99, seed=36, noise=0):
         Parameter for the logistic map. Values beyond 3.56995 exhibit
         chaotic behaviour.
     seed : int
-        Sets the random seed for the logistic map.
-    noise : float
+        Sets the random seed for numpy's random number generator.
+    eta : float
         Amplitude of noise to add to the logistic map.
 
     Returns
     -------
     X : 1D array
         Logistic map of size (sz).
+
+    References
+    ----------
+    __ [1] : Wikipedia, "Logistic Map", https://en.wikipedia.org/wiki/Logistic_map.
 
     """
 
@@ -48,15 +55,19 @@ def logistic_map(sz=256, A=3.99, seed=36, noise=0):
         # logistic equation in space
         X[tt+1] = A*X[tt]*(1-X[tt])
 
-    X += noise*np.random.rand(sz)
+    X += eta*np.random.rand(sz)
 
     return X
 
-def noisy_periodic(sz=256,freq=52,seed=36,noise=.5):
+def noisy_periodic(sz=256, freq=52, seed=36, eta=.5):
     """A simple periodic equation with a a specified amplitude of noise.
 
-    t = linspace(0, freq*pi, sz)
-    X = sin(t) + .5cos(t) + noise
+    .. math::
+
+      X(t) = sin( 2\pi ft ) + 0.5cos( 2\pi ft ) + \eta
+
+    Where :math:`f` is the frequency and :math:`\eta` is the amplitude of the
+    noise.
 
     Parameters
     ----------
@@ -65,8 +76,8 @@ def noisy_periodic(sz=256,freq=52,seed=36,noise=.5):
     freq : int
         Frequency of the periodic equation.
     seed : int
-        Sets the random seed for reproducible results.
-    noise : float
+        Sets the random seed for numpy's random number generator.
+    eta : float
         Amplitude of the noise.
 
     Returns
@@ -78,60 +89,25 @@ def noisy_periodic(sz=256,freq=52,seed=36,noise=.5):
 
     np.random.seed(seed=seed)
 
-    t = np.linspace(0, freq*np.pi, sz)  #prep range for averages
+    t = np.linspace(0, freq, sz)  #prep range for averages
 
-    X = np.sin(t) + .5*np.cos(t) + noise * np.random.rand(sz)
+    X = np.sin(2*np.pi*t) + .5*np.cos(2*np.pi*t) + eta * np.random.rand(sz)
 
     #all positive and between 0 and 1
     X = X + np.abs(np.min(X))
     X = X/np.max(X)
     return X
 
-def noisy_periodic_complicated(sz=256, freq=52, seed=36, noise=.5):
[…]
