+++ docs/generate-data.rst
@@ -1,17 +1,19 @@
 Generate Data
 =============
 
-skedm comes with both 1D and 2D test data. These built in functions allow you to quickly explore the behavior of different systems.
+skedm comes with both 1D and 2D test data. These built in functions allow you to quickly explore the behavior of different systems. Check the `notebook`_ for visual examples.
 
 1 Dimensional
 ^^^^^^^^^^^^^
 
 .. automodule:: skedm.data
-   :members: logistic_map, noisy_periodic, noisy_periodic_complicated, noise_1D, lorenz
+   :members: logistic_map, noisy_periodic, noise_1d, lorenz
 
 
 2 Dimensional
 ^^^^^^^^^^^^^
 
 .. automodule:: skedm.data
-   :members: chaos_2d, periodic_2d, periodic_brown, brown_noise, noise_2d, chaos_3d, random_circles, circle_in_circle, circles_with_stuff, random_sized_circles, voronoi_matrix
+   :members: chaos_2d, periodic_2d, brown_noise, periodic_brown, noise_2d, overlapping_circles, concentric_circles, small_and_large_circles, random_sized_circles, voronoi_matrix, chaos_3d
+
+.. _notebook: https://github.com/nickc1/skedm/tree/master/scripts
