### 2017-03-29 2c0855e
added links to docstrings
