+++ skedm/data.py
@@ -5,20 +5,19 @@
 
 import numpy as np
 from numpy import genfromtxt
-#from scipy import integrate
 from sklearn import neighbors
 import scipy.ndimage
 
 
-def logistic_map(sz=256, A=3.99, seed=36, eta=0):
-    """Solutions to the logistic map for a given amount of time steps [1]_.
+def logistic_map(sz=256, A=3.99, seed=36, noise=0):
+    """Solutions to the `logistic map`_ for a given amount of time steps.
 
     .. math::
 
-      X_{t+1} = AX_t(1 - X_t) + \eta
+      X_{t+1} = AX_t(1 - X_t) + \\alpha\\eta
 
-    where :math:`A` is the parameter that controls chaos, and :math:`\eta` is
-    the amplitude of the noise.
+    where :math:`A` is the parameter that controls chaos, :math:`\eta` is
+    uncorrelated white noise and :math:`\\alpha` is the amplitude of the noise.
 
     Parameters
     ----------
@@ -29,17 +28,16 @@ def logistic_map(sz=256, A=3.99, seed=36, eta=0):
         chaotic behaviour.
     seed : int
         Sets the random seed for numpy's random number generator.
-    eta : float
-        Amplitude of noise to add to the logistic map.
+    noise : float
+        Amplitude of the noise.
 
     Returns
     -------
     X : 1D array
         Logistic map of size (sz).
 
-    References
-    ----------
-    __ [1] : Wikipedia, "Logistic Map", https://en.wikipedia.org/wiki/Logistic_map.
+
+    .. _logistic map : https://en.wikipedia.org/wiki/Logistic_map
 
     """
 
@@ -55,19 +53,19 @@ def logistic_map(sz=256, A=3.99, seed=36, eta=0):
         # logistic equation in space
         X[tt+1] = A*X[tt]*(1-X[tt])
 
-    X += eta*np.random.rand(sz)
+    X += noise*np.random.rand(sz)
 
     return X
 
-def noisy_periodic(sz=256, freq=52, seed=36, eta=.5):
+def noisy_periodic(sz=256, freq=52, noise=.5, seed=36):
     """A simple periodic equation with a a specified amplitude of noise.
 
     .. math::
 
-      X(t) = sin( 2\pi ft ) + 0.5cos( 2\pi ft ) + \eta
+      X(t) = sin( 2\pi ft ) + 0.5cos( 2\pi ft ) + \\alpha\\eta
 
-    Where :math:`f` is the frequency and :math:`\eta` is the amplitude of the
-    noise.
+    Where :math:`f` is the frequency :math:`\eta` is uncorrelated white noise
+    and :math:`\\alpha` is the amplitude of the noise.
 
     Parameters
     ----------
@@ -75,10 +73,10 @@ def noisy_periodic(sz=256, freq=52, seed=36, eta=.5):
         Length of the time series.
     freq : int
         Frequency of the periodic equation.
+    noise : float
+        Amplitude of the noise.
     seed : int
         Sets the random seed for numpy's random number generator.
-    eta : float
-        Amplitude of the noise.
 
     Returns
     -------
@@ -91,7 +89,7 @@ def noisy_periodic(sz=256, freq=52, seed=36, eta=.5):
 
     t = np.linspace(0, freq, sz)  #prep range for averages
 
-    X = np.sin(2*np.pi*t) + .5*np.cos(2*np.pi*t) + eta * np.random.rand(sz)
+    X = np.sin(2*np.pi*t) + .5*np.cos(2*np.pi*t) + noise * np.random.rand(sz)
 
     #all positive and between 0 and 1
     X = X + np.abs(np.min(X))
@@ -122,8 +120,8 @@ def noise_1d(sz=256, seed=36):
     return X
 
 
[…]
