### 2017-03-29 948f884
fixed some typos in docstrings
