### 2017-04-14 70923a2
added the tarball and added some dependencies to environment.yml
