### 2017-04-14 #207 @nickc1
Hey John,

I think I have it all sorted out. Again, thanks for taking the time to come through my documentation.

> Version: Does the release version given match the GitHub release (v0.0.1)?

This should be all set now. Both are version v0.0.3. I did a fresh pip install in a conda environment and everything seems to run properly.

> I think you mean y is the green matrix above (right?).

Fixed!

> Documentation: Better, however there is still some confusing/incomplete documentation:

Fixed! Thanks for calling me out on blobber/blobber2. Those were ridiculous names. I also added a notebook (skedm_example_data.ipynb) giving examples of all of the test data.

> For the Lorenz attractor, if you have time/it's easy, it might make sense to let the user change the default sigma, beta, and rho values.

Check.

> For the chaos_2d and chaos_3d -- could you provide a link with more information about the logistic map diffused in space? Or if it's simple to give an explicit definition that would be helpful too.

Explicit definition given.

> Same with periodic_2d (i.e. what is the explicit form of the equation?)

Explicit form given.

> In periodic_brown, the equaiton given is X = sin(y + noise*2pi), however there's a freq param; I'm assuming it's X = sin(2 * pi * (freq * y + noise))?

Added to docstring.

> For skedm.data.noise_2d(sz=128, seed=36); I imagine some people will want to know what library/algorithm you're using as the random number generation; I would just mention what library/function you're using (e.g. numpy, etc.).

Added this to all functions utilizing numpy's random number generation.

> random_circles -- what is a "random gaussian circle?"

Updated docstring with links explaining what is going on.

> circle_in_circle, circles_with_stuff,random_sized_circles, I'm not sure what the form of the output is. What does the blobber/blobber2 function do? If you're returning A 2-D array of size (sz, sz), what does that array represent? Is it a bit-map image? The explanation for these should be a bit more specific

Updated docstring and added this to the notebook.

> In the voronoi_diagram, you say sz -- row and column size of the space in which the circle is placed., but I think that's a typo since these are polygons. The return value is also unclear to me (see previous note)

Fixed

Let me know if you need me to make any changes!

Thanks again,
Nick
