### 2017-04-17 #207 @jakevdp
Thanks for the response, Nick. @johnh2o2 – please let us know if the responses address all your concerns. Thanks!
