### 2017-04-17 #207 @johnh2o2
Hi @nickc1,

Sorry about the delay! This looks great. I feel comfortable recommending this for acceptance in JOSS as is.

John
