### 2017-04-17 #207 @jakevdp
Thanks @nickc1 and @johnh2o2! Looks good – we'll conclude the review process shortly (cc @arfon).
