### 2017-04-19 4312f4c
updated typo in docs
