### 2017-04-19 #207 @arfon
@johnh2o2 - many thanks for your review here and @jakevdp for editing this submission ✨

@nickc1 - your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00207 ⚡️ 🚀 💥
