### 2017-04-19 #207 @nickc1
Thanks again @johnh2o2 for all the thoughtful comments. The package and documentation are way more user friendly. Thanks @jakevdp for editing. This was a cool submission process and definitely learned quite a bit along the way. Thanks for taking the time to make skedm better!
