### 2017-05-24 #207 @nickc1
@johnh2o2 @jakevdp Realized I inadvertently left off the citation that actually coined the term empirical dynamic modeling. Any way to add it in in the first sentence after the first mention of the term? or is it all locked in?

Thanks!

Ye, Hao, et al. "Equation-free mechanistic ecosystem forecasting using empirical dynamic modeling." Proceedings of the National Academy of Sciences 112.13 (2015): E1569-E1576.
